source/auth/requirements.txt
astroid==2.12.13
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==5.2.0
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==2.6.0
PyJWT==2.6.0
pylint==2.15.9
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
